Refactor home page to improve structure and user experience by modularizing components. Introduce Header, FolderSelection, AnalysisOptions, ProgressIndicator, and ImageResults components. Update image analysis handling and download logic for better clarity and functionality.
